fix K task

+++ Algo/contest1/contest1taskK/k.c
@@ -50,12 +50,12 @@ int partition(int* arr, const int low, const int high) {
     return i;
 }
 
-void InsertionSort(int* arr, const int low, const int high) {
-    for (int i = low + 1; i <= high; i++) {
-        int key = arr[i];
-        int j = i - 1;
+void InsertionSort(int* arr, const size_t low, const size_t high) {
+    for (size_t i = low + 1; i <= high; i++) {
+        size_t key = arr[i];
+        size_t j = i - 1;
 
-        while (j >= low && arr[j] > key) {
+        while (j >= low && (size_t)arr[j] > key) {
             arr[j + 1] = arr[j];
             j--;
         }
